pyproject.toml
[tool.poetry]
name = "toolchemy"
version = "0.2.185"
description = "a set of auxiliary programming tools"
authors = ["Cyprian Nosek <cyprian.nosek@protonmail.com>"]

[tool.poetry.dependencies]
python = "~3.12"
ollama = "^0.6.0"
torch = "^2.7.0"
numpy = "2.0.2"
mlflow = "^3.8.1"
neptune-scale = "^0.13.0"
openai = "^1.82.1"
pillow = "^11.2.1"
wyoming = "^1.7.1"
python-dateutil = "^2.9.0.post0"
tinydb = "^4.8.2"
tenacity = "^9.1.2"
diskcache = "^5.6.3"
google-genai = "^1.55.0"
colorlog = "^6.10.1"
jsonschema = "^4.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pyfakefs = "^5.9.2"
pytest-benchmark = "^5.1.0"
freezegun = "^1.5.5"
deepdiff = "^8.6.1"

[tool.poetry.scripts]
prompt-studio = "toolchemy.ai.prompter:run_studio"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
